+++ netbox/core/utils.py
@@ -1,3 +1,4 @@
+from django.db import DatabaseError, connection
 from django.http import Http404
 from django.utils.translation import gettext_lazy as _
 from django_rq.queues import get_queue, get_queue_by_index, get_redis_connection
@@ -18,6 +19,7 @@ from rq.registry import (
 __all__ = (
     'delete_rq_job',
     'enqueue_rq_job',
+    'get_db_schema',
     'get_rq_jobs',
     'get_rq_jobs_from_status',
     'requeue_rq_job',
@@ -154,3 +156,50 @@ def stop_rq_job(job_id):
     queue = get_queue_by_index(queue_index)
 
     return stop_jobs(queue, job_id)[0]
+
+
+def get_db_schema():
+    """
+    Query the current PostgreSQL schema and return a list of tables, each with its columns and
+    indexes. Returns an empty list if the database is not accessible.
+    """
+    db_schema = []
+    try:
+        with connection.cursor() as cursor:
+            cursor.execute("""
+                SELECT table_name, column_name, data_type, is_nullable, column_default
+                FROM information_schema.columns
+                WHERE table_schema = current_schema()
+                ORDER BY table_name, ordinal_position
+            """)
+            columns_by_table = {}
+            for table_name, column_name, data_type, is_nullable, column_default in cursor.fetchall():
+                columns_by_table.setdefault(table_name, []).append({
+                    'name': column_name,
+                    'type': data_type,
+                    'nullable': is_nullable == 'YES',
+                    'default': column_default,
+                })
+
+            cursor.execute("""
+                SELECT tablename, indexname, indexdef
+                FROM pg_indexes
+                WHERE schemaname = current_schema()
+                ORDER BY tablename, indexname
+            """)
+            indexes_by_table = {}
+            for table_name, index_name, index_def in cursor.fetchall():
+                indexes_by_table.setdefault(table_name, []).append({
+                    'name': index_name,
+                    'definition': index_def,
+                })
+
+        for table_name in sorted(columns_by_table.keys()):
+            db_schema.append({
+                'name': table_name,
+                'columns': columns_by_table[table_name],
+                'indexes': indexes_by_table.get(table_name, []),
+            })
+    except DatabaseError:
+        pass
+    return db_schema
